move resolution and unify imports
